requirements.txt
aiohappyeyeballs==2.4.4 ; python_version >= "3.13" and python_version < "4.0"
aiohttp==3.11.11 ; python_version >= "3.13" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.13" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.13" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.13" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.13" and python_version < "4.0"
attrs==24.3.0 ; python_version >= "3.13" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.13" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.13" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.13" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.13" and python_version < "4.0"
cohere==5.13.6 ; python_version >= "3.13" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4.0" and platform_system == "Windows"
dataclasses-json==0.6.7 ; python_version >= "3.13" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.13" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.13" and python_version < "4.0"
fastavro==1.10.0 ; python_version >= "3.13" and python_version < "4.0"
filelock==3.16.1 ; python_version >= "3.13" and python_version < "4.0"
flask-cors==5.0.0 ; python_version >= "3.13" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.13" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.13" and python_version < "4.0"
fsspec==2024.12.0 ; python_version >= "3.13" and python_version < "4.0"
gotrue==2.11.1 ; python_version >= "3.13" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.13"
groq==0.15.0 ; python_version >= "3.13" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.13" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.13" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.13" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.13" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.13" and python_version < "4.0"
httpx-sse==0.4.0 ; python_version >= "3.13" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.13" and python_version < "4.0"
httpx[http2]==0.27.2 ; python_version >= "3.13" and python_version < "4.0"
huggingface-hub==0.27.1 ; python_version >= "3.13" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.13" and python_version < "4.0"
idna==3.10 ; python_version >= "3.13" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.13" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.13" and python_version < "4.0"
jq==1.8.0 ; python_version >= "3.13" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.13" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.13" and python_version < "4.0"
langchain-cohere==0.3.4 ; python_version >= "3.13" and python_version < "4.0"
langchain-community==0.3.14 ; python_version >= "3.13" and python_version < "4.0"
langchain-core==0.3.29 ; python_version >= "3.13" and python_version < "4.0"
langchain-experimental==0.3.4 ; python_version >= "3.13" and python_version < "4.0"
langchain-groq==0.2.3 ; python_version >= "3.13" and python_version < "4.0"
langchain-text-splitters==0.3.5 ; python_version >= "3.13" and python_version < "4.0"
langchain==0.3.14 ; python_version >= "3.13" and python_version < "4.0"
langsmith==0.2.10 ; python_version >= "3.13" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.13" and python_version < "4.0"
marshmallow==3.25.1 ; python_version >= "3.13" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.13" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.13" and python_version < "4.0"
numpy==2.2.1 ; python_version >= "3.13" and python_version < "4.0"
orjson==3.10.14 ; python_version >= "3.13" and python_version < "4.0" and platform_python_implementation != "PyPy"
packaging==24.2 ; python_version >= "3.13" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.13" and python_version < "4.0"
parameterized==0.9.0 ; python_version >= "3.13" and python_version < "4.0"
postgrest==0.19.1 ; python_version >= "3.13" and python_version < "4.0"
propcache==0.2.1 ; python_version >= "3.13" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.13" and python_version < "4.0"
pydantic-settings==2.7.1 ; python_version >= "3.13" and python_version < "4.0"
pydantic==2.10.5 ; python_version >= "3.13" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.13" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.13" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.13" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.13" and python_version < "4.0"
rank-bm25==0.2.2 ; python_version >= "3.13" and python_version < "4.0"
realtime==2.1.0 ; python_version >= "3.13" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.13" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.13" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.13" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.13" and python_version < "4.0"
sqlalchemy==2.0.37 ; python_version >= "3.13" and python_version < "4.0"
storage3==0.11.0 ; python_version >= "3.13" and python_version < "4.0"
strenum==0.4.15 ; python_version >= "3.13" and python_version < "4.0"
supabase==2.11.0 ; python_version >= "3.13" and python_version < "4.0"
supafunc==0.9.0 ; python_version >= "3.13" and python_version < "4.0"
tabulate==0.9.0 ; python_version >= "3.13" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.13" and python_version < "4.0"
tokenizers==0.21.0 ; python_version >= "3.13" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.13" and python_version < "4.0"
types-requests==2.32.0.20241016 ; python_version >= "3.13" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.13" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.13" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.13" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.13" and python_version < "4.0"
websockets==13.1 ; python_version >= "3.13" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.13" and python_version < "4.0"
yarl==1.18.3 ; python_version >= "3.13" and python_version < "4.0"
